Module aws_sdk_comprehendmedical::error

Modules

internal_server_exception

See InternalServerException

invalid_encoding_exception

See InvalidEncodingException

invalid_request_exception

See InvalidRequestException

resource_not_found_exception

See ResourceNotFoundException

service_unavailable_exception

See ServiceUnavailableException

text_size_limit_exceeded_exception

See TextSizeLimitExceededException

too_many_requests_exception

See TooManyRequestsException

validation_exception

See ValidationException

Structs

DescribeEntitiesDetectionV2JobError

DescribeICD10CMInferenceJobError

DescribePHIDetectionJobError

DescribeRxNormInferenceJobError

DetectEntitiesError

DetectEntitiesV2Error

DetectPHIError

InferICD10CMError

InferRxNormError

InternalServerException

An internal server error occurred. Retry your request.

InvalidEncodingException

The input text was not in valid UTF-8 character encoding. Check your text then retry your

InvalidRequestException

The request that you made is invalid. Check your request to determine why it's invalid

ListEntitiesDetectionV2JobsError

ListICD10CMInferenceJobsError

ListPHIDetectionJobsError

ListRxNormInferenceJobsError

ResourceNotFoundException

The resource identified by the specified Amazon Resource Name (ARN) was not found. Check

ServiceUnavailableException

The Amazon Comprehend Medical service is temporarily unavailable. Please wait and then retry your request.

StartEntitiesDetectionV2JobError

StartICD10CMInferenceJobError

StartPHIDetectionJobError

StartRxNormInferenceJobError

StopEntitiesDetectionV2JobError

StopICD10CMInferenceJobError

StopPHIDetectionJobError

StopRxNormInferenceJobError

TextSizeLimitExceededException

The size of the text you submitted exceeds the size limit. Reduce the size of the text or

TooManyRequestsException

You have made too many requests within a short period of time. Wait for a short time and

ValidationException

The filter that you specified for the operation is invalid. Check the filter values that

Enums

DescribeEntitiesDetectionV2JobErrorKind

DescribeICD10CMInferenceJobErrorKind

DescribePHIDetectionJobErrorKind

DescribeRxNormInferenceJobErrorKind

DetectEntitiesErrorKind

DetectEntitiesV2ErrorKind

DetectPHIErrorKind

InferICD10CMErrorKind

InferRxNormErrorKind

ListEntitiesDetectionV2JobsErrorKind

ListICD10CMInferenceJobsErrorKind

ListPHIDetectionJobsErrorKind

ListRxNormInferenceJobsErrorKind

StartEntitiesDetectionV2JobErrorKind

StartICD10CMInferenceJobErrorKind

StartPHIDetectionJobErrorKind

StartRxNormInferenceJobErrorKind

StopEntitiesDetectionV2JobErrorKind

StopICD10CMInferenceJobErrorKind

StopPHIDetectionJobErrorKind

StopRxNormInferenceJobErrorKind