Module aws_sdk_comprehendmedical::model

Modules

attribute

See Attribute

comprehend_medical_async_job_filter

See ComprehendMedicalAsyncJobFilter

comprehend_medical_async_job_properties

See ComprehendMedicalAsyncJobProperties

entity

See Entity

icd10_cm_attribute

See Icd10CmAttribute

icd10_cm_concept

See Icd10CmConcept

icd10_cm_entity

See Icd10CmEntity

icd10_cm_trait

See Icd10CmTrait

input_data_config

See InputDataConfig

output_data_config

See OutputDataConfig

rx_norm_attribute

See RxNormAttribute

rx_norm_concept

See RxNormConcept

rx_norm_entity

See RxNormEntity

rx_norm_trait

See RxNormTrait

trait

See Trait

unmapped_attribute

See UnmappedAttribute

Structs

Attribute

An extracted segment of the text that is an attribute of an entity, or otherwise related

ComprehendMedicalAsyncJobFilter

Provides information for filtering a list of detection jobs.

ComprehendMedicalAsyncJobProperties

Provides information about a detection job.

Entity

Provides information about an extracted medical entity.

Icd10CmAttribute

The detected attributes that relate to an entity. This includes an extracted segment of

Icd10CmConcept

The ICD-10-CM concepts that the entity could refer to, along with a score indicating the

Icd10CmEntity

The collection of medical entities extracted from the input text and their associated

Icd10CmTrait

Contextual information for the entity. The traits recognized by InferICD10CM are

InputDataConfig

The input properties for an entities detection job. This includes the name of the S3

OutputDataConfig

The output properties for a detection job.

RxNormAttribute

The extracted attributes that relate to this entity. The attributes recognized by

RxNormConcept

The RxNorm concept that the entity could refer to, along with a score indicating the

RxNormEntity

The collection of medical entities extracted from the input text and their associated

RxNormTrait

The contextual information for the entity. InferRxNorm recognizes the trait

Trait

Provides contextual information about the extracted entity.

UnmappedAttribute

An attribute that we extracted, but were unable to relate to an entity.

Enums

AttributeName

EntitySubType

EntityType

Icd10CmAttributeType

Icd10CmEntityCategory

Icd10CmEntityType

Icd10CmRelationshipType

Icd10CmTraitName

JobStatus

LanguageCode

RelationshipType

RxNormAttributeType

RxNormEntityCategory

RxNormEntityType

RxNormTraitName