Module aws_sdk_comprehendmedical::operation

Structs

DescribeEntitiesDetectionV2Job

Gets the properties associated with a medical entities detection job. Use this operation

DescribeICD10CMInferenceJob

Gets the properties associated with an InferICD10CM job. Use this operation to get the

DescribePHIDetectionJob

Gets the properties associated with a protected health information (PHI) detection job.

DescribeRxNormInferenceJob

Gets the properties associated with an InferRxNorm job. Use this operation to get the

DetectEntities

The DetectEntities operation is deprecated. You should use the DetectEntitiesV2 operation instead.

DetectEntitiesV2

Inspects the clinical text for a variety of medical entities and returns specific

DetectPHI

Inspects the clinical text for protected health information (PHI) entities and returns

InferICD10CM

InferICD10CM detects medical conditions as entities listed in a patient record and links

InferRxNorm

InferRxNorm detects medications as entities listed in a patient record and links to the

ListEntitiesDetectionV2Jobs

Gets a list of medical entity detection jobs that you have submitted.

ListICD10CMInferenceJobs

Gets a list of InferICD10CM jobs that you have submitted.

ListPHIDetectionJobs

Gets a list of protected health information (PHI) detection jobs that you have

ListRxNormInferenceJobs

Gets a list of InferRxNorm jobs that you have submitted.

StartEntitiesDetectionV2Job

Starts an asynchronous medical entity detection job for a collection of documents. Use the

StartICD10CMInferenceJob

Starts an asynchronous job to detect medical conditions and link them to the ICD-10-CM

StartPHIDetectionJob

Starts an asynchronous job to detect protected health information (PHI). Use the

StartRxNormInferenceJob

Starts an asynchronous job to detect medication entities and link them to the RxNorm

StopEntitiesDetectionV2Job

Stops a medical entities detection job in progress.

StopICD10CMInferenceJob

Stops an InferICD10CM inference job in progress.

StopPHIDetectionJob

Stops a protected health information (PHI) detection job in progress.

StopRxNormInferenceJob

Stops an InferRxNorm inference job in progress.